split formatter into several files
